Icon component adjustments, get rid of Layout component
